RSA demo

+++ knapsack2.cpp
@@ -2,45 +2,45 @@
 #include <math.h>
 using namespace std;
 /*
-int max(int x, int y) {
+float max(float x, float y) {
    return (x > y) ? x : y;
 }
 */
-void knapSack(int W, int w[], int v[], int n) {
-   int i, wt;
-   int K[n + 1][W + 1];
+void knapSack(float W, float w[], float v[], int n) {
+   float i, wt;
+   float K[n + 1][(int)W + 1];
    for (i = 0; i <= n; i++) {
       for (wt = 0; wt <= W; wt++) {
          if (i == 0 || wt == 0)
-         K[i][wt] = 0;
-         else if (w[i - 1] <= wt)
-            K[i][wt] = max(v[i - 1] + K[i - 1][wt - w[i - 1]], K[i - 1][wt]);
+         K[(int)i][(int)wt] = 0;
+         else if (w[(int)i - 1] <= wt)
+            K[(int)i][(int)wt] = max(v[(int)i - 1] + K[(int)i - 1][(int)(wt - w[(int)i - 1])], K[(int)i - 1][(int)wt]);
          else
-        K[i][wt] = K[i - 1][wt];
+        K[(int)i][(int)wt] = K[(int)i - 1][(int)wt];
       }
    }
-   int cw = W, cn = n;// cw là khối lượng còn lại, cn là số vật còn lại
+   float cw = W, cn = n;// cw là khối lượng còn lại, cn là số vật còn lại
    cout<<"Item was selected: ";
    while(cw != 0 && cn != 0){
-      if(K[cn][cw] != K[cn-1][cw] ){
+      if(K[(int)cn][(int)cw] != K[(int)cn-1][cw] ){
          cout<<cn<<" ";
-         cw = cw - w[cn-1];
+         cw = cw - w[(int)cn-1];
       }
       cn--;
    }
    cout<<"\n";
    cout <<"Value of items was selected: "<< K[n][W];
 }
-int main() {
+float main() {
    cout << "Enter the number of items in a Knapsack:";
-   int n, W;
+   float n, W;
    cin >> n;
    cout << "Enter the capacity of knapsack: ";
    cin >> W;
-   int v[n], w[n];
-   for (int i = 0; i < n; i++) {
+   float v[(int)n], w[(int)n];
+   for (float i = 0; i < n; i++) {
       cout << "Enter weight and value for item " << i+1 << ":";
-      cin >> w[i];
+      cin >> w[(int)i];
       cin >> v[i];
    }
    knapSack(W, w, v, n);
